[winutil] refactor GetSidFromUser to avoid dep cycle
