+++ pkg/secrets/check_rights_windows.go
@@ -14,7 +14,6 @@ import (
 
 	"golang.org/x/sys/windows"
 
-	"github.com/DataDog/datadog-agent/pkg/api/security"
 	"github.com/DataDog/datadog-agent/pkg/util/winutil"
 )
 
@@ -80,7 +79,7 @@ func checkRights(filename string, allowGroupExec bool) error {
 	}
 	defer windows.FreeSid(administrators)
 
-	secretuser, err := security.GetSidFromUser()
+	secretuser, err := winutil.GetSidFromUser()
 	defer windows.FreeSid(secretuser)
 
 	bSecretUserExplicitlyAllowed := false
